feat: Verify & Fix Numerical Correctness across Backends
- Implemented  suite covering Auction, Sinkhorn, Dual Ascent.
- [Dual Ascent] Switched Python/Triton backends to stable Gradient Ascent (fixing Newton instability).
- [Auction] Added BF16 dispatch to CUDA kernels.
- [Docs] Added  detailing Tensor Core activation strategy.
- [Compiled] Optimized  with standard inductor options.

+++ src/efficient_linear_assignment/auction/triton_backend.py
@@ -598,7 +598,9 @@ class AuctionTriton:
         benefits = -cost_matrix
         
         B, N, M = benefits.shape
-        prices = torch.zeros((B, M), device=benefits.device, dtype=torch.float32)
+        # User Request: Momory Transfer in Float16 (if input is F16)
+        # We store prices in same dtype as benefits.
+        prices = torch.zeros((B, M), device=benefits.device, dtype=benefits.dtype)
         assignment = torch.full((B, N), -1, device=benefits.device, dtype=torch.long)
         
         # Pre-allocate buffers
